[refactor] comment domain
